将Sparse从16bit量化回滚到32bit量化并通过RTL仿真

+++ CSIM/Sparse_HLS/Sparse_HLS/src/sparse.hpp
@@ -23,5 +23,5 @@ void sparse(
 	t_AXI_DataType *inputs,
 	t_AXI_DataType *outputs,
 	int shift,
-	int16_t mul,
+	int mul,
 	bool &sparse_flag);
